+++ DrainLib/Engines/UnknownEngineAccessor.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 namespace DrainLib.Engines {
 	public class UnknownEngineAccessor : BaseEngineAccessor {
-		internal UnknownEngineAccessor(ScummVMConnector connector, uint engineAddr) : base(connector, engineAddr) {
+		internal UnknownEngineAccessor(ScummVMConnector connector, IntPtr engineAddr) : base(connector, engineAddr) {
 		}
 
 		public override string GameId => "???";
